added transaction and payment management system with cart functionality, checkout process, order management for buyers and sellers, transaction tracking, payment method CRUD, and comprehensive filtering/search capabilities

+++ database/migrations/2025_12_25_100002_create_products_table.php
@@ -18,8 +18,8 @@ return new class extends Migration
             $table->string('name_en')->nullable();
             $table->string('slug')->unique();
             $table->longText('description');
-            $table->decimal('price', 10, 2);
-            $table->decimal('compare_price', 10, 2)->nullable(); // Strikethrough price
+            $table->unsignedBigInteger('price'); // Store in cents (paisa)
+            $table->unsignedBigInteger('compare_price')->nullable(); // Strikethrough price in cents
             $table->unsignedInteger('stock_count')->default(0);
             $table->unsignedInteger('stock_alert_threshold')->default(5);
             $table->string('sku')->nullable();
